impl(sidekick): Simplified help implementation
